Small refactoring + doc fixes
