+++ pyrow/__init__.py
@@ -1,0 +1,5 @@
+"""
+PyRow
+
+PyRow is a Python library that allows interaction with Concept2 PM3, PM4 or PM5.
+"""
