+++ pyrow/const.py
@@ -1,1 +1,3 @@
+"""pyrow constants."""
+
 __version__ = '0.1.0'
